refactor: refactor merge nodes & merge edges
