+++ graphgen/operators/build_kg/build_kg.py
@@ -5,22 +5,19 @@ import gradio as gr
 
 from graphgen.bases.base_storage import BaseGraphStorage
 from graphgen.bases.datatypes import Chunk
-from graphgen.models import LightRAGKGBuilder, OpenAIClient, Tokenizer
-from graphgen.operators.build_kg.merge_kg import merge_edges, merge_nodes
+from graphgen.models import LightRAGKGBuilder, OpenAIClient
 from graphgen.utils import run_concurrent
 
 
 async def build_kg(
     llm_client: OpenAIClient,
     kg_instance: BaseGraphStorage,
-    tokenizer_instance: Tokenizer,
     chunks: List[Chunk],
     progress_bar: gr.Progress = None,
 ):
     """
     :param llm_client: Synthesizer LLM model to extract entities and relationships
     :param kg_instance
-    :param tokenizer_instance
     :param chunks
     :param progress_bar: Gradio progress bar to show the progress of the extraction
     :return:
@@ -44,7 +41,16 @@ async def build_kg(
         for k, v in e.items():
             edges[tuple(sorted(k))].extend(v)
 
-    await merge_nodes(nodes, kg_instance, llm_client, tokenizer_instance)
-    await merge_edges(edges, kg_instance, llm_client, tokenizer_instance)
+    await run_concurrent(
+        lambda kv: kg_builder.merge_nodes(kv, kg_instance=kg_instance),
+        list(nodes.items()),
+        desc="Inserting entities into storage",
+    )
+
+    await run_concurrent(
+        lambda kv: kg_builder.merge_edges(kv, kg_instance=kg_instance),
+        list(edges.items()),
+        desc="Inserting relationships into storage",
+    )
 
     return kg_instance
